feat: search by category

+++ src/main/java/dev/fousin/movieflix/controller/MovieController.java
@@ -6,6 +6,7 @@ import dev.fousin.movieflix.entity.Movie;
 import dev.fousin.movieflix.mapper.MovieMapper;
 import dev.fousin.movieflix.service.MovieService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -15,9 +16,9 @@ import java.util.Optional;
 
 @RestController
 @RequestMapping("/movieflix/movie")
-@RequiredArgsConstructor
 public class MovieController {
-    private final MovieService service;
+    @Autowired
+    private MovieService service;
 
     @PostMapping
     public ResponseEntity<MovieResponse> save(@RequestBody MovieRequest request) {
@@ -32,6 +33,12 @@ public class MovieController {
         return ResponseEntity.ok(service.findAll());
     }
 
+    @GetMapping("/search")
+    public ResponseEntity<List<MovieResponse>> getByCategory(@RequestParam Long category) {
+        return ResponseEntity.ok(service.getByCategory(category));
+    }
+
+
     @GetMapping("/{id}")
     public ResponseEntity<MovieResponse> getById(@PathVariable Long id) {
         return service.getById(id)
@@ -41,15 +48,19 @@ public class MovieController {
 
     }
 
-//    @PutMapping("/{id}")
-//    public ResponseEntity<MovieResponse> update(@PathVariable Long id, @RequestBody MovieRequest request) {
-//        Movie movie = MovieMapper.toEntity(request);
-//        MovieResponse response = MovieMapper.toResponse(service.update(id, MovieMapper.toEntity(request)));
-//        return response;
-//    }
+    @PutMapping("/{id}")
+    public ResponseEntity<MovieResponse> update(@PathVariable Long id, @RequestBody MovieRequest request) {
+        return service.update(id, MovieMapper.toEntity(request))
+                .map( movie -> ResponseEntity.ok(MovieMapper.toResponse(movie)))
+                .orElse(ResponseEntity.notFound().build());
+    }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> delete(@PathVariable Long id) {
+        Optional<Movie> opt = service.getById(id);
+        if(!opt.isPresent()) {
+            return ResponseEntity.notFound().build();
+        }
         service.delete(id);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
